refactor: move password recovery workflow into its own blueprint
Separate the password recovery workflow from the login blueprint where it was originally implemented.
