+++ authserver/api/__init__.py
@@ -5,3 +5,4 @@ from authserver.api.oauth2 import oauth2_bp
 from authserver.api.role import role_bp
 from authserver.api.home import home_bp
 from authserver.api.scope import scope_bp
+from authserver.api.password_recovery import password_recovery_bp
