+++ authserver/api/templates/login.html
@@ -8,7 +8,7 @@
         {{ render_field(form.username, class='form-control', autofocus='true') }}
         {{ render_field(form.password, class='form-control') }}
         <div class="forgot-password-link">
-            <a href={{ url_for('home_ep.recover_password') }}>Forgot Password?</a>
+            <a href={{ url_for('password_recovery_ep.recover_password') }}>Forgot Password?</a>
         </div>
         <button class="mt-5 btn btn-primary btn-lg" type="submit">Log-in</button>
     </form>
